# infra/openapi-dotnet.yaml
openapi: 3.0.1
info:
  title: GitHub Issues API
  version: v1
servers:
  - url: https://localhost:5051
paths:
  /issues:
    get:
      tags:
        - Issues
      operationId: Issues
      parameters:
        - name: User
          in: query
          required: true
          style: form
          schema:
            type: string
        - name: Repository
          in: query
          required: true
          style: form
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitHubIssueCollectionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /issues/{id}:
    get:
      tags:
        - Issues
      operationId: IssueById
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
        - name: User
          in: query
          required: true
          style: form
          schema:
            type: string
        - name: Repository
          in: query
          required: true
          style: form
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitHubIssueItemResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          nullable: true
      additionalProperties: false
    GitHubIssueCollectionResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GitHubIssueItemResponse'
          nullable: true
      additionalProperties: false
    GitHubIssueItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
          nullable: true
        number:
          type: integer
          format: int32
          nullable: true
        title:
          type: string
          nullable: true
        body:
          type: string
          nullable: true
      additionalProperties: false
  securitySchemes:
    github_token:
      type: apiKey
      description: Please enter valid GitHub Token
      name: x-github-token
      in: header
    api_key:
      type: apiKey
      description: Please enter valid API Key
      name: x-webapi-key
      in: header
security:
  - github_token: []
  - api_key: []